chore: Move selectors to ts file
