+++ ui/selectors/selectors.js
@@ -6,7 +6,6 @@ import {
   getLocalizedSnapManifest,
   SnapStatus,
 } from '@metamask/snaps-utils';
-import { BigNumber } from 'bignumber.js';
 import { memoize } from 'lodash';
 import semver from 'semver';
 import { createSelector } from 'reselect';
@@ -91,7 +90,6 @@ import {
   isAddressLedger,
   getIsUnlocked,
   getCompletedOnboarding,
-  getTokenBalances,
 } from '../ducks/metamask/metamask';
 import {
   getLedgerWebHidConnectedStatus,
@@ -111,8 +109,6 @@ import { hasTransactionData } from '../../shared/modules/transaction.utils';
 import { toChecksumHexAddress } from '../../shared/modules/hexstring-utils';
 import { createDeepEqualSelector } from '../../shared/modules/selectors/util';
 import { isSnapIgnoredInProd } from '../helpers/utils/snaps';
-import { calculateTokenBalance } from '../components/app/assets/util/calculateTokenBalance';
-import { calculateTokenFiatAmount } from '../components/app/assets/util/calculateTokenFiatAmount';
 import {
   getAllUnapprovedTransactions,
   getCurrentNetworkTransactions,
@@ -131,7 +127,6 @@ import {
   getMultichainBalances,
   getMultichainNetworkProviders,
 } from './multichain';
-import { getAccountAssets, getAssetsMetadata, getAssetsRates } from './assets';
 
 /** `appState` slice selectors */
 
@@ -2985,152 +2980,3 @@ export function getKeyringSnapAccounts(state) {
   return keyringAccounts;
 }
 ///: END:ONLY_INCLUDE_IF
-
-export const getTokenBalancesEvm = createDeepEqualSelector(
-  getSelectedAccountTokensAcrossChains,
-  getSelectedAccountNativeTokenCachedBalanceByChainId,
-  getTokenBalances,
-  (_state, accountAddress) => accountAddress,
-  getMarketData,
-  getCurrencyRates,
-  getPreferences,
-  getIsTokenNetworkFilterEqualCurrentNetwork,
-  getCurrentNetwork,
-  (
-    selectedAccountTokensChains,
-    nativeBalances,
-    tokenBalances,
-    selectedAccountAddress,
-    marketData,
-    currencyRates,
-    preferences,
-    isOnCurrentNetwork,
-    currentNetwork,
-  ) => {
-    const { hideZeroBalanceTokens } = preferences;
-    const selectedAccountTokenBalancesAcrossChains =
-      tokenBalances[selectedAccountAddress];
-
-    // we need to filter Testnets
-    const isTestNetwork = TEST_CHAINS.includes(currentNetwork.chainId);
-    const filteredAccountTokensChains = Object.fromEntries(
-      Object.entries(selectedAccountTokensChains).filter(([chainId]) =>
-        isTestNetwork
-          ? TEST_CHAINS.includes(chainId)
-          : !TEST_CHAINS.includes(chainId),
-      ),
-    );
-    const tokensWithBalance = [];
-    Object.entries(filteredAccountTokensChains).forEach(
-      ([stringChainKey, tokens]) => {
-        const chainId = stringChainKey;
-        tokens.forEach((token) => {
-          const { isNative, address, decimals } = token;
-          const balance =
-            calculateTokenBalance({
-              isNative,
-              chainId,
-              address,
-              decimals,
-              nativeBalances,
-              selectedAccountTokenBalancesAcrossChains,
-            }) || '0';
-
-          const tokenFiatAmount = calculateTokenFiatAmount({
-            token,
-            chainId,
-            balance,
-            marketData,
-            currencyRates,
-          });
-
-          // Respect the "hide zero balance" setting (when true):
-          // - Native tokens should always display with zero balance when on the current network filter.
-          // - Native tokens should not display with zero balance when on all networks filter
-          // - ERC20 tokens with zero balances should respect the setting on both the current and all networks.
-
-          // Respect the "hide zero balance" setting (when false):
-          // - Native tokens should always display with zero balance when on the current network filter.
-          // - Native tokens should always display with zero balance when on all networks filter
-          // - ERC20 tokens always display with zero balance on both the current and all networks filter.
-          if (
-            !hideZeroBalanceTokens ||
-            balance !== '0' ||
-            (token.isNative && isOnCurrentNetwork)
-          ) {
-            tokensWithBalance.push({
-              ...token,
-              balance,
-              tokenFiatAmount,
-              chainId,
-              string: String(balance),
-            });
-          }
-        });
-      },
-    );
-    return tokensWithBalance;
-  },
-);
-
-export const getMultiChainAssets = createDeepEqualSelector(
-  (_state, selectedAccount) => selectedAccount,
-  getMultichainBalances,
-  getAccountAssets,
-  getAssetsMetadata,
-  getAssetsRates,
-  (_state) => _state.metamask.conversionRates,
-  (
-    selectedAccountAddress,
-    multichainBalances,
-    accountAssets,
-    assetsMetadata,
-    assetRates,
-    multichainCoinRates,
-  ) => {
-    const assetIds = accountAssets?.[selectedAccountAddress.id] || [];
-    const balances = multichainBalances?.[selectedAccountAddress.id];
-    return assetIds.map((assetId) => {
-      const [chainId, assetDetails] = assetId.split('/');
-      const isNative = assetDetails.split(':')[0] === 'slip44';
-      const balance = balances?.[assetId] || { amount: '0', unit: '' };
-      const rate = assetRates?.[assetId]?.rate || '0';
-      const nativeRate = multichainCoinRates?.[assetId]?.rate || '0';
-      const balanceInFiat = new BigNumber(balance.amount).times(rate);
-      const nativeBalanceInFiat = new BigNumber(balance.amount).times(
-        nativeRate,
-      );
-      const metadata = assetsMetadata[assetId] || {
-        name: balance.unit,
-        symbol: balance.unit || '',
-        fungible: true,
-        units: [{ name: assetId, symbol: balance.unit || '', decimals: 0 }],
-      };
-      // not super happy with this override here
-      let tokenImage = '';
-      let secondary;
-      if (isNative) {
-        tokenImage = CHAIN_ID_TOKEN_IMAGE_MAP[chainId] || '';
-        secondary = nativeBalanceInFiat.toNumber();
-      } else {
-        tokenImage = metadata.iconUrl || '';
-        secondary = balanceInFiat.toNumber();
-      }
-      const decimals = metadata.units[0]?.decimals || 0;
-      return {
-        title: isNative ? balance.unit : metadata.name,
-        address: assetId,
-        symbol: metadata.symbol,
-        image: tokenImage,
-        decimals,
-        chainId,
-        isNative,
-        primary: balance.amount,
-        secondary,
-        string: '',
-        tokenFiatAmount: balanceInFiat, // for now we are keeping this is to satisfy sort, this should be fiat amount
-        isStakeable: false,
-      };
-    });
-  },
-);
